app/artist-model-service/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30
pytest ~= 5.4
psycopg2==2.9.9
grpcio-health-checking ~= 1.62.1